% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document_primer.R
\name{html_document_primer}
\alias{html_document_primer}
\title{Rmarkdown HTML Document based on Primer CSS}
\usage{
html_document_primer(
  ...,
  auto_theme = TRUE,
  light_theme = "light",
  dark_theme = "dark_dimmed",
  list_style = NULL,
  enable_checkboxes = FALSE,
  css = NULL,
  toc = FALSE,
  toc_depth = 3,
  header = TRUE,
  mathjax = NULL,
  fig_width = 10,
  fig_height = 7,
  fig_retina = 2,
  footer = TRUE,
  keep_md = FALSE,
  dev = "png",
  highlight = TRUE,
  pandoc_args = NULL,
  extra_dependencies = NULL,
  md_extensions = NULL,
  self_contained = TRUE
)
}
\arguments{
\item{auto_theme}{Indicates whether the theme should be changed based on
user's system preferences}

\item{light_theme}{The light theme, e.g., "light"}

\item{dark_theme}{The dark theme, e.g., "dark", "dark_dimmed",
"dark_high_contrast"}

\item{list_style}{Indicates whether list elements should use bullets
or not}

\item{enable_checkboxes}{Indicates whether or not check boxes can be modified}

\item{css}{CSS and/or Sass files to include. Files with an extension of .sass
or .scss are compiled to CSS via \code{sass::sass()}. Also, if \code{theme} is a
\code{\link[bslib:bs_theme]{bslib::bs_theme()}} object, Sass code may reference the relevant Bootstrap
Sass variables, functions, mixins, etc.}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{header}{Indicates whether the title should be included as a
header in the output document}

\item{mathjax}{Include mathjax. The "default" option uses an https URL from a
MathJax CDN. The "local" option uses a local version of MathJax (which is
copied into the output directory). You can pass an alternate URL or pass
\code{NULL} to exclude MathJax entirely.}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{fig_retina}{Scaling to perform for retina displays (defaults to 2, which
currently works for all widely used retina displays). Set to \code{NULL} to
prevent retina scaling. Note that this will always be \code{NULL} when
\code{keep_md} is specified (this is because \code{fig_retina} relies on
outputting HTML directly into the markdown document).}

\item{footer}{Indicates whether the footer should be shown}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{highlight}{Indicates whether or not the code should be highlighted}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link[rmarkdown]{html_document_base}}}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition of R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
its size).}
}
\value{
An R Markdown output format that can be used with \verb{output:} in an
\code{.Rmd} or for use with \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
HTML documents with R Markdown.
}
\section{MathJax}{
 Note that MathJax is disabled by default to reduce the
overall size of the final document. You can enable MathJax by setting
\code{mathjax = "default"}, see \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}} for more options.
}

\examples{
html_document_primer()

\dontrun{
rmarkdown::render("input.Rmd", html_document_primer())
}

}
