% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primermd-package.R
\docType{package}
\name{primermd}
\alias{primermd}
\alias{primermd-package}
\title{primermd}
\description{
This is a R Markdown theme based on GitHub's Primer CSS. It
supports both light/dark theme.
}
\details{
This is a R Markdown theme based on GitHub's Primer CSS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://primermd.amirmasoudabdol.name}
  \item Report bugs at \url{https://github.com/amirmasoudabdol/primermd/issues}
}

}
\author{
Amir Masoud Abdol (i@amirmasoudabdol.name)
}
\keyword{internal}
